Remove `ImageUploader` component and integrate inline file inputs
The `ImageUploader` component was removed in favor of implementing simpler file input handling inline within the product pages. This streamlines the codebase and improves maintainability while still supporting image uploading with necessary validations and previews.
